Moved drill down function to a separate utils file.
